refactor: replace MonacoCelCDN with MonacoCelEditor and clean up unused files
