+++ keep-ui/shared/ui/MonacoCELEditor/index.ts
@@ -1,12 +1,1 @@
-"use client";
-
-import dynamic from "next/dynamic";
-
-const MonacoCel = dynamic(
-  () => import("./MonacoCelNPM").then((mod) => mod.MonacoCelNPM),
-  {
-    ssr: false,
-  }
-);
-
-export { MonacoCel };
+export { MonacoCelEditor } from "./monaco-cel-editor";
