--- keep-ui/shared/ui/MonacoCELEditor/index.turbopack.ts
@@ -1,3 +1,0 @@
-// This file is used to replace the default export for the MonacoEditor component
-// when using Turbopack in development mode
-export { MonacoCelCDN as MonacoCel } from "./MonacoCelCDN";
