+++ keep-ui/shared/ui/MonacoCELEditor/monaco-cel-base.turbopack.tsx
@@ -1,0 +1,39 @@
+"use client";
+
+import { Editor, EditorProps, loader } from "@monaco-editor/react";
+import { useEffect, useRef, useState } from "react";
+
+interface MonacoCelProps extends EditorProps {
+  onMonacoLoaded?: (monacoInstance: typeof import("monaco-editor")) => void;
+  onMonacoLoadFailure?: (error: Error) => void;
+}
+
+export function MonacoCel(props: MonacoCelProps) {
+  const [isLoaded, setIsLoaded] = useState(false);
+  const onMonacoLoadedRef = useRef<MonacoCelProps["onMonacoLoaded"] | null>(
+    null
+  );
+  onMonacoLoadedRef.current = props.onMonacoLoaded;
+  const onMonacoLoadFailureRef = useRef<
+    MonacoCelProps["onMonacoLoadFailure"] | null
+  >();
+  onMonacoLoadFailureRef.current = props.onMonacoLoadFailure;
+
+  useEffect(() => {
+    loader
+      .init()
+      .then((monacoInstance) => {
+        onMonacoLoadedRef.current?.(monacoInstance);
+        setIsLoaded(true);
+      })
+      .catch((error: Error) => {
+        onMonacoLoadFailureRef.current?.(error);
+      });
+  }, []);
+
+  if (!isLoaded) {
+    return null;
+  }
+
+  return <Editor {...props} />;
+}
